updated tuning code and analysis, started cleanup
