+++ config.py
@@ -35,15 +35,15 @@ import os
 
 import matplotlib as mpl
 #some basic text parameters for figures
-#mpl.rcParams['font.family'] = "Atkinson Hyperlegible" # if installed but not showing up, rebuild mpl cache
-mpl.rcParams['font.size'] = 18
+mpl.rcParams['font.family'] = "Atkinson Hyperlegible" # if installed but not showing up, rebuild mpl cache
+mpl.rcParams['font.size'] = 10
 mpl.rcParams['savefig.format'] = 'pdf'
 mpl.rcParams['axes.unicode_minus'] = False
-mpl.rcParams['axes.titlesize'] = 32
-mpl.rcParams['axes.labelsize'] = 24
+# mpl.rcParams['axes.titlesize'] = 14
+# mpl.rcParams['axes.labelsize'] = 12
 mpl.rcParams['axes.titlelocation'] = 'center'
 mpl.rcParams['axes.titleweight'] = 'bold'
 mpl.rcParams['figure.constrained_layout.use'] = True
-mpl.rcParams['figure.titlesize'] = 32
+# mpl.rcParams['figure.titlesize'] = 14
 mpl.rcParams['figure.titleweight'] = 'bold'
 mpl.rcParams['pdf.fonttype'] = 42
